etc/requirements.txt
async-timeout==5.0.1
attrs==25.4.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
croniter==6.0.0
exceptiongroup==1.3.0
flasgger==0.9.7.1
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
hiredis==3.3.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.2
mistune==3.1.4
packaging==25.0
pluggy==1.6.0
prettytable==3.16.0
psycopg2==2.9.10
Pygments==2.19.2
PyMySQL==1.1.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
requests==2.32.4
rpds-py==0.27.1
six==1.17.0
soupsieve==2.7
tomli==2.2.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
